experiments/requirements.txt
mlflow
scikit-learn
python-dotenv
boto3
prometheus_client
joblib


